@@ -3,13 +3,19 @@
 // ----------------------------------------------------------------------------
 
 import * as language from "./Language";
-import * as tle from "./TLE";
 
 /**
  * A completion item in the list of completion suggestions that appear when a user invokes auto-completion (Ctrl + Space).
  */
 export class Item {
-    constructor(private _name: string, private _insertText: string, private _insertSpan: language.Span, private _detail: string, private _description: string, private _type: CompletionKind) {
+    constructor(
+        private _name: string,
+        private _insertText: string,
+        private _insertSpan: language.Span,
+        private _detail: string,
+        private _description: string,
+        private _type: CompletionKind
+    ) {
     }
 
     public get name(): string {
