@@ -7,7 +7,7 @@
 // tslint:disable:function-name // Grandfathered in
 // tslint:disable:cyclomatic-complexity // Grandfathered in
 
-import * as assert from "assert";
+import { assert } from "./fixed_assert";
 
 import * as language from "./Language";
 import * as basic from "./Tokenizer";
@@ -99,7 +99,7 @@ export function RightCurlyBracket(startIndex: number): Token {
     return new Token(TokenType.RightCurlyBracket, startIndex, [basic.RightCurlyBracket]);
 }
 
-export function LeftSquareBracket(startIndex: number) {
+export function LeftSquareBracket(startIndex: number): Token {
     return new Token(TokenType.LeftSquareBracket, startIndex, [basic.LeftSquareBracket]);
 }
 
@@ -255,7 +255,9 @@ export function readNumber(iterator: utilities.Iterator<basic.Token>): basic.Tok
             numberBasicTokens.push(iterator.current());
             iterator.moveNext();
 
-            if (iterator.current() && (iterator.current().getType() === basic.TokenType.Dash || iterator.current().getType() === basic.TokenType.Plus)) {
+            if (iterator.current()
+                && (iterator.current().getType() === basic.TokenType.Dash || iterator.current().getType() === basic.TokenType.Plus)
+            ) {
                 // Exponent number sign
                 numberBasicTokens.push(iterator.current());
                 iterator.moveNext();
@@ -329,9 +331,9 @@ export class Tokenizer {
         // assign the functions to be "hasStarted: this.hasStarted", then the function get assigned
         // without a bound "this" pointer.
         return {
-            hasStarted: () => { return this.hasStarted(); },
-            current: () => { return this.currentBasicToken(); },
-            moveNext: () => { return this.moveNextBasicToken(); }
+            hasStarted: (): boolean => { return this.hasStarted(); },
+            current: (): basic.Token => { return this.currentBasicToken(); },
+            moveNext: (): boolean => { return this.moveNextBasicToken(); }
         };
     }
 
@@ -390,7 +392,10 @@ export class Tokenizer {
                         switch (this.currentBasicTokenType()) {
                             case basic.TokenType.ForwardSlash:
                                 const lineCommentBasicTokens: basic.Token[] = [basic.ForwardSlash, basic.ForwardSlash];
-                                while (this.moveNextBasicToken() && this.currentBasicTokenType() !== basic.TokenType.NewLine && this.currentBasicTokenType() !== basic.TokenType.CarriageReturnNewLine) {
+                                while (this.moveNextBasicToken()
+                                    && this.currentBasicTokenType() !== basic.TokenType.NewLine
+                                    && this.currentBasicTokenType() !== basic.TokenType.CarriageReturnNewLine
+                                ) {
                                     lineCommentBasicTokens.push(this.currentBasicToken());
                                 }
                                 this._current = Comment(this._currentTokenStartIndex, lineCommentBasicTokens);
@@ -567,6 +572,7 @@ export class ObjectValue extends Value {
      * stack. If the provided property name stack is empty, then return this value.
      */
     public getPropertyValueFromStack(propertyNameStack: string[]): Value {
+        // tslint:disable-next-line:no-this-assignment
         let result: Value = this;
 
         while (result && propertyNameStack.length > 0) {
@@ -597,7 +603,7 @@ export class ObjectValue extends Value {
  * A property in a JSON ObjectValue.
  */
 export class Property extends Value {
-    constructor(span: language.Span, private _name: StringValue, private _value: Value) {
+    constructor(span: language.Span, private _name: StringValue, private _value: Value | null) {
         super(span);
     }
 
@@ -615,7 +621,7 @@ export class Property extends Value {
     /**
      * The value of the property.
      */
-    public get value(): Value {
+    public get value(): Value | null {
         return this._value;
     }
 
@@ -825,9 +831,13 @@ export class ParseResult {
      * line and column indexes.
      */
     public getCharacterIndex(lineIndex: number, columnIndex: number): number {
+        // tslint:disable-next-line:max-line-length
         assert(0 <= lineIndex, `Cannot get a character index for a negative line index (${lineIndex}).`);
+        // tslint:disable-next-line:max-line-length
         assert(lineIndex < this.lineLengths.length, `Cannot get a character index for a line index greater than the number of parsed lines (lineIndex: ${lineIndex}, lines parsed: ${this.lineLengths.length}).`);
+        // tslint:disable-next-line:max-line-length
         assert(0 <= columnIndex, `Cannot get a character index for a negative columnIndex (${columnIndex}).`);
+        // tslint:disable-next-line:max-line-length
         assert(columnIndex <= this.getMaxColumnIndex(lineIndex), `Cannot get a character index for a columnIndex (${columnIndex}) that is greater than the lineIndex's (${lineIndex}) line max column index (${this.getMaxColumnIndex(lineIndex)}).`);
 
         let characterIndex = columnIndex;
@@ -871,6 +881,7 @@ export class ParseResult {
     }
 
     private getToken(tokenIndex: number): Token {
+        // tslint:disable-next-line:max-line-length
         assert(0 <= tokenIndex && tokenIndex < this.tokenCount, `The tokenIndex (${tokenIndex}) must always be between 0 and the token count - 1 (${this.tokenCount - 1}).`);
 
         return this._tokens[tokenIndex];
@@ -981,7 +992,7 @@ export function parse(stringValue: string): ParseResult {
  * All of the Tokens that are read will be placed into the provided
  * tokens array.
  */
-function parseValue(tokenizer: Tokenizer, tokens: Token[]): Value {
+function parseValue(tokenizer: Tokenizer, tokens: Token[]): Value | null {
     let value: Value = null;
 
     if (!tokenizer.hasStarted()) {
@@ -1055,9 +1066,11 @@ function parseObject(tokenizer: Tokenizer, tokens: Token[]): ObjectValue {
                 propertyName = null;
             }
         } else {
-            const propertyValue: Value = parseValue(tokenizer, tokens);
-            propertySpan = propertySpan.union(propertyValue.span);
-            objectSpan = objectSpan.union(propertyValue.span);
+            const propertyValue: Value | null = parseValue(tokenizer, tokens);
+            if (propertyValue) {
+                propertySpan = propertySpan.union(propertyValue.span);
+                objectSpan = objectSpan.union(propertyValue.span);
+            }
 
             properties.push(new Property(propertySpan, propertyName, propertyValue));
 
