@@ -3,17 +3,15 @@
 // ----------------------------------------------------------------------------
 
 import * as assert from "assert";
-
-import * as Completion from "../src/Completion";
-import * as language from "../src/Language";
+import { Completion, Language } from "../extension.bundle";
 
 suite("Completion", () => {
     suite("Item", () => {
         test("constructor(string, Span, string, string, Type)", () => {
-            const item: Completion.Item = new Completion.Item("a", "b", new language.Span(1, 2), "c", "d", Completion.CompletionKind.Function);
+            const item: Completion.Item = new Completion.Item("a", "b", new Language.Span(1, 2), "c", "d", Completion.CompletionKind.Function);
             assert.deepStrictEqual(item.description, "d");
             assert.deepStrictEqual(item.detail, "c");
-            assert.deepStrictEqual(item.insertSpan, new language.Span(1, 2));
+            assert.deepStrictEqual(item.insertSpan, new Language.Span(1, 2));
             assert.deepStrictEqual(item.insertText, "b");
             assert.deepStrictEqual(item.name, "a");
             assert.deepStrictEqual(item.kind, Completion.CompletionKind.Function);
