@@ -2,22 +2,14 @@
 // Copyright (c) Microsoft Corporation.  All rights reserved.
 // ----------------------------------------------------------------------------
 
-// tslint:disable:no-unused-expression
-// tslint:disable:max-func-body-length
-// tslint:disable:promise-function-async
+// tslint:disable:no-unused-expression max-func-body-length promise-function-async max-line-length insecure-random
+// tslint:disable:object-literal-key-quotes
 
 import * as assert from "assert";
-
-import * as Json from "../src/JSON";
-import * as language from "../src/Language";
-import * as Reference from "../src/Reference";
-
-import { DeploymentTemplate } from "../src/DeploymentTemplate";
-import { ReferenceInVariableDefinitionJSONVisitor } from "../src/DeploymentTemplate";
-import { Histogram } from "../src/Histogram";
-import { IncorrectArgumentsCountIssue } from "../src/IncorrectArgumentsCountIssue";
-import { ParameterDefinition } from "../src/ParameterDefinition";
-import { UnrecognizedFunctionIssue } from "../src/UnrecognizedFunctionIssue";
+import { randomBytes } from "crypto";
+import { ISuiteCallbackContext } from "mocha";
+import { DeploymentTemplate, Histogram, IncorrectArgumentsCountIssue, Json, Language, ParameterDefinition, Reference, ReferenceInVariableDefinitionJSONVisitor, UnrecognizedFunctionIssue } from "../extension.bundle";
+import { testDiagnostics } from "./testDiagnostics";
 
 suite("DeploymentTemplate", () => {
     suite("constructor(string)", () => {
@@ -72,7 +64,7 @@ suite("DeploymentTemplate", () => {
             assert(pd0);
             assert.deepStrictEqual(pd0.name.toString(), "num");
             assert.deepStrictEqual(pd0.description, null);
-            assert.deepStrictEqual(pd0.span, new language.Span(18, 27));
+            assert.deepStrictEqual(pd0.span, new Language.Span(18, 27));
             assert.equal(null, dt.schemaUri);
         });
 
@@ -86,7 +78,7 @@ suite("DeploymentTemplate", () => {
             assert(pd0);
             assert.deepStrictEqual(pd0.name.toString(), "num");
             assert.deepStrictEqual(pd0.description, null);
-            assert.deepStrictEqual(pd0.span, new language.Span(18, 64));
+            assert.deepStrictEqual(pd0.span, new Language.Span(18, 64));
             assert.equal(null, dt.schemaUri);
         });
 
@@ -100,7 +92,7 @@ suite("DeploymentTemplate", () => {
             assert(pd0);
             assert.deepStrictEqual(pd0.name.toString(), "num");
             assert.deepStrictEqual(pd0.description, "");
-            assert.deepStrictEqual(pd0.span, new language.Span(18, 62));
+            assert.deepStrictEqual(pd0.span, new Language.Span(18, 62));
             assert.equal(null, dt.schemaUri);
         });
 
@@ -114,7 +106,7 @@ suite("DeploymentTemplate", () => {
             assert(pd0);
             assert.deepStrictEqual(pd0.name.toString(), "num");
             assert.deepStrictEqual(pd0.description, "num description");
-            assert.deepStrictEqual(pd0.span, new language.Span(18, 77));
+            assert.deepStrictEqual(pd0.span, new Language.Span(18, 77));
             assert.equal(null, dt.schemaUri);
         });
 
@@ -135,7 +127,7 @@ suite("DeploymentTemplate", () => {
 
             const variableDefinition: Json.StringValue = Json.asStringValue(dt.variableDefinitions[0].value);
             assert(variableDefinition);
-            assert.deepStrictEqual(variableDefinition.span, new language.Span(22, 3));
+            assert.deepStrictEqual(variableDefinition.span, new Language.Span(22, 3));
             assert.deepStrictEqual(variableDefinition.toString(), "A");
 
             assert.equal(null, dt.schemaUri);
@@ -150,13 +142,13 @@ suite("DeploymentTemplate", () => {
             assert.deepStrictEqual(dt.variableDefinitions[0].name.toString(), "a");
             const a: Json.StringValue = Json.asStringValue(dt.variableDefinitions[0].value);
             assert(a);
-            assert.deepStrictEqual(a.span, new language.Span(22, 3));
+            assert.deepStrictEqual(a.span, new Language.Span(22, 3));
             assert.deepStrictEqual(a.toString(), "A");
 
             assert.deepStrictEqual(dt.variableDefinitions[1].name.toString(), "b");
             const b: Json.NumberValue = Json.asNumberValue(dt.variableDefinitions[1].value);
             assert(b);
-            assert.deepStrictEqual(b.span, new language.Span(32, 1));
+            assert.deepStrictEqual(b.span, new Language.Span(32, 1));
 
             assert.equal(null, dt.schemaUri);
         });
@@ -199,21 +191,29 @@ suite("DeploymentTemplate", () => {
     suite("errors", () => {
         test("with empty deployment template", () => {
             const dt = new DeploymentTemplate("", "id");
-            return dt.errors.then((errors: language.Issue[]) => {
+            return dt.errors.then((errors: Language.Issue[]) => {
                 assert.deepStrictEqual(errors, []);
             });
         });
 
+        testDiagnostics(
+            "no errors: empty",
+            {
+            },
+            {},
+            [
+            ]);
+
         test("with empty object deployment template", () => {
             const dt = new DeploymentTemplate("{}", "id");
-            return dt.errors.then((errors: language.Issue[]) => {
+            return dt.errors.then((errors: Language.Issue[]) => {
                 assert.deepStrictEqual(errors, []);
             });
         });
 
         test("with one property deployment template", () => {
             const dt = new DeploymentTemplate("{ 'name': 'value' }", "id");
-            return dt.errors.then((errors: language.Issue[]) => {
+            return dt.errors.then((errors: Language.Issue[]) => {
                 assert.deepStrictEqual(errors, []);
             });
         });
@@ -221,9 +221,9 @@ suite("DeploymentTemplate", () => {
         test("with one TLE parse error deployment template", () => {
             const dt = new DeploymentTemplate("{ 'name': '[concat()' }", "id");
             const expectedErrors = [
-                new language.Issue(new language.Span(20, 1), "Expected a right square bracket (']').")
+                new Language.Issue(new Language.Span(20, 1), "Expected a right square bracket (']').")
             ];
-            return dt.errors.then((errors: language.Issue[]) => {
+            return dt.errors.then((errors: Language.Issue[]) => {
                 assert.deepStrictEqual(errors, expectedErrors);
             });
         });
@@ -231,9 +231,9 @@ suite("DeploymentTemplate", () => {
         test("with one undefined parameter error deployment template", () => {
             const dt = new DeploymentTemplate("{ 'name': '[parameters(\"test\")]' }", "id");
             const expectedErrors = [
-                new language.Issue(new language.Span(23, 6), "Undefined parameter reference: \"test\"")
+                new Language.Issue(new Language.Span(23, 6), "Undefined parameter reference: \"test\"")
             ];
-            return dt.errors.then((errors: language.Issue[]) => {
+            return dt.errors.then((errors: Language.Issue[]) => {
                 assert.deepStrictEqual(errors, expectedErrors);
             });
         });
@@ -241,9 +241,9 @@ suite("DeploymentTemplate", () => {
         test("with one undefined variable error deployment template", () => {
             const dt = new DeploymentTemplate("{ 'name': '[variables(\"test\")]' }", "id");
             const expectedErrors = [
-                new language.Issue(new language.Span(22, 6), "Undefined variable reference: \"test\"")
+                new Language.Issue(new Language.Span(22, 6), "Undefined variable reference: \"test\"")
             ];
-            return dt.errors.then((errors: language.Issue[]) => {
+            return dt.errors.then((errors: Language.Issue[]) => {
                 assert.deepStrictEqual(errors, expectedErrors);
             });
         });
@@ -251,83 +251,83 @@ suite("DeploymentTemplate", () => {
         test("with one unrecognized function error deployment template", () => {
             const dt = new DeploymentTemplate("{ 'name': '[blah(\"test\")]' }", "id");
             const expectedErrors = [
-                new UnrecognizedFunctionIssue(new language.Span(12, 4), "blah")
+                new UnrecognizedFunctionIssue(new Language.Span(12, 4), "blah")
             ];
-            return dt.errors.then((errors: language.Issue[]) => {
+            return dt.errors.then((errors: Language.Issue[]) => {
                 assert.deepStrictEqual(errors, expectedErrors);
             });
         });
 
         test("with reference() call in variable definition", () => {
             const dt = new DeploymentTemplate(`{ "variables": { "a": "[reference('test')]" } }`, "id");
-            return dt.errors.then((errors: language.Issue[]) => {
+            return dt.errors.then((errors: Language.Issue[]) => {
                 assert.deepStrictEqual(
                     errors,
-                    [new language.Issue(new language.Span(24, 9), "reference() cannot be invoked inside of a variable definition.")]
+                    [new Language.Issue(new Language.Span(24, 9), "reference() cannot be invoked inside of a variable definition.")]
                 );
             });
         });
 
         test("with reference() call inside a different expression in a variable definition", () => {
             const dt = new DeploymentTemplate(`{ "variables": { "a": "[concat(reference('test'))]" } }`, "id");
-            return dt.errors.then((errors: language.Issue[]) => {
+            return dt.errors.then((errors: Language.Issue[]) => {
                 assert.deepStrictEqual(
                     errors,
-                    [new language.Issue(new language.Span(31, 9), "reference() cannot be invoked inside of a variable definition.")]);
+                    [new Language.Issue(new Language.Span(31, 9), "reference() cannot be invoked inside of a variable definition.")]);
             });
         });
 
         test("with unnamed property access on variable reference", () => {
             const dt = new DeploymentTemplate(`{ "variables": { "a": {} }, "z": "[variables('a').]" }`, "id");
-            return dt.errors.then((errors: language.Issue[]) => {
+            return dt.errors.then((errors: Language.Issue[]) => {
                 assert.deepStrictEqual(
                     errors,
-                    [new language.Issue(new language.Span(50, 1), "Expected a literal value.")]);
+                    [new Language.Issue(new Language.Span(50, 1), "Expected a literal value.")]);
             });
         });
 
         test("with property access on variable reference without variable name", () => {
             const dt = new DeploymentTemplate(`{ "variables": { "a": {} }, "z": "[variables().b]" }`, "id");
-            return dt.errors.then((errors: language.Issue[]) => {
+            return dt.errors.then((errors: Language.Issue[]) => {
                 assert.deepStrictEqual(
                     errors,
-                    [new IncorrectArgumentsCountIssue(new language.Span(35, 11), "The function 'variables' takes 1 argument.", "variables", 0, 1, 1)]);
+                    [new IncorrectArgumentsCountIssue(new Language.Span(35, 11), "The function 'variables' takes 1 argument.", "variables", 0, 1, 1)]);
             });
         });
 
         test("with property access on string variable reference", () => {
             const dt = new DeploymentTemplate(`{ "variables": { "a": "A" }, "z": "[variables('a').b]" }`, "id");
-            return dt.errors.then((errors: language.Issue[]) => {
+            return dt.errors.then((errors: Language.Issue[]) => {
                 assert.deepStrictEqual(
                     errors,
-                    [new language.Issue(new language.Span(51, 1), `Property "b" is not a defined property of "variables('a')".`)]);
+                    [new Language.Issue(new Language.Span(51, 1), `Property "b" is not a defined property of "variables('a')".`)]);
             });
         });
 
         test("with undefined variable reference child property", () => {
             const dt = new DeploymentTemplate(`{ "variables": { "a": {} }, "z": "[variables('a').b]" }`, "id");
-            return dt.errors.then((errors: language.Issue[]) => {
+            return dt.errors.then((errors: Language.Issue[]) => {
                 assert.deepStrictEqual(
                     errors,
-                    [new language.Issue(new language.Span(50, 1), `Property "b" is not a defined property of "variables('a')".`)]);
+                    [new Language.Issue(new Language.Span(50, 1), `Property "b" is not a defined property of "variables('a')".`)]);
             });
         });
 
         test("with undefined variable reference grandchild property", () => {
             const dt = new DeploymentTemplate(`{ "variables": { "a": { "b": {} } }, "z": "[variables('a').b.c]" }`, "id");
-            return dt.errors.then((errors: language.Issue[]) => {
+            return dt.errors.then((errors: Language.Issue[]) => {
                 assert.deepStrictEqual(
                     errors,
-                    [new language.Issue(new language.Span(61, 1), `Property "c" is not a defined property of "variables('a').b".`)]);
+                    [new Language.Issue(new Language.Span(61, 1), `Property "c" is not a defined property of "variables('a').b".`)]);
             });
         });
 
         test("with undefined variable reference child and grandchild properties", () => {
             const dt = new DeploymentTemplate(`{ "variables": { "a": { "d": {} } }, "z": "[variables('a').b.c]" }`, "id");
-            return dt.errors.then((errors: language.Issue[]) => {
+            return dt.errors.then((errors: Language.Issue[]) => {
                 assert.deepStrictEqual(
                     errors,
-                    [new language.Issue(new language.Span(59, 1), `Property "b" is not a defined property of "variables('a')".`)]);
+                    [new Language.Issue(new Language.Span(59, 1), `Property "b" is not a defined property of "variables('a')".`)]);
             });
         });
     });
@@ -337,7 +337,7 @@ suite("DeploymentTemplate", () => {
             const dt = new DeploymentTemplate(`{ "parameters": { "a": {} } }`, "id");
             assert.deepStrictEqual(
                 dt.warnings,
-                [new language.Issue(new language.Span(18, 3), "The parameter 'a' is never used.")]);
+                [new Language.Issue(new Language.Span(18, 3), "The parameter 'a' is never used.")]);
         });
 
         test("with no unused parameters", () => {
@@ -350,7 +350,7 @@ suite("DeploymentTemplate", () => {
             const dt = new DeploymentTemplate(`{ "variables": { "a": "A" } }`, "id");
             assert.deepStrictEqual(
                 dt.warnings,
-                [new language.Issue(new language.Span(17, 3), "The variable 'a' is never used.")]);
+                [new Language.Issue(new Language.Span(17, 3), "The variable 'a' is never used.")]);
         });
 
         test("with no unused variables", () => {
@@ -456,7 +456,7 @@ suite("DeploymentTemplate", () => {
             assert(pd0);
             assert.deepStrictEqual(pd0.name.toString(), "a");
             assert.deepStrictEqual(pd0.description, null);
-            assert.deepStrictEqual(pd0.span, new language.Span(18, 7));
+            assert.deepStrictEqual(pd0.span, new Language.Span(18, 7));
         });
 
         test("with parameter with metadata but no description", () => {
@@ -468,7 +468,7 @@ suite("DeploymentTemplate", () => {
             assert(pd0);
             assert.deepStrictEqual(pd0.name.toString(), "a");
             assert.deepStrictEqual(pd0.description, null);
-            assert.deepStrictEqual(pd0.span, new language.Span(18, 23));
+            assert.deepStrictEqual(pd0.span, new Language.Span(18, 23));
         });
 
         test("with parameter with metadata and description", () => {
@@ -480,7 +480,7 @@ suite("DeploymentTemplate", () => {
             assert(pd0);
             assert.deepStrictEqual(pd0.name.toString(), "a");
             assert.deepStrictEqual(pd0.description, "b");
-            assert.deepStrictEqual(pd0.span, new language.Span(18, 43));
+            assert.deepStrictEqual(pd0.span, new Language.Span(18, 43));
         });
     });
 
@@ -526,9 +526,9 @@ suite("DeploymentTemplate", () => {
             assert(apples);
             assert.deepStrictEqual(apples.name.toString(), "apples");
             assert.deepStrictEqual(apples.description, null);
-            assert.deepStrictEqual(apples.span, new language.Span(18, 30));
-            assert.deepStrictEqual(apples.name.span, new language.Span(18, 8), "Wrong name.span");
-            assert.deepStrictEqual(apples.name.unquotedSpan, new language.Span(19, 6), "Wrong name.unquotedSpan");
+            assert.deepStrictEqual(apples.span, new Language.Span(18, 30));
+            assert.deepStrictEqual(apples.name.span, new Language.Span(18, 8), "Wrong name.span");
+            assert.deepStrictEqual(apples.name.unquotedSpan, new Language.Span(19, 6), "Wrong name.unquotedSpan");
         });
 
         test("with one-sided-quote match", () => {
@@ -537,7 +537,7 @@ suite("DeploymentTemplate", () => {
             assert(apples);
             assert.deepStrictEqual(apples.name.toString(), "apples");
             assert.deepStrictEqual(apples.description, null);
-            assert.deepStrictEqual(apples.span, new language.Span(18, 30));
+            assert.deepStrictEqual(apples.span, new Language.Span(18, 30));
         });
 
         test("with quoted match", () => {
@@ -546,7 +546,7 @@ suite("DeploymentTemplate", () => {
             assert(apples);
             assert.deepStrictEqual(apples.name.toString(), "apples");
             assert.deepStrictEqual(apples.description, null);
-            assert.deepStrictEqual(apples.span, new language.Span(18, 30));
+            assert.deepStrictEqual(apples.span, new Language.Span(18, 30));
         });
 
         test("with case insensitive match", () => {
@@ -555,7 +555,7 @@ suite("DeploymentTemplate", () => {
             assert(apples);
             assert.deepStrictEqual(apples.name.toString(), "apples");
             assert.deepStrictEqual(apples.description, null);
-            assert.deepStrictEqual(apples.span, new language.Span(18, 30));
+            assert.deepStrictEqual(apples.span, new Language.Span(18, 30));
         });
 
         test("with case sensitive and insensitive match", () => {
@@ -564,13 +564,13 @@ suite("DeploymentTemplate", () => {
             assert(APPLES);
             assert.deepStrictEqual(APPLES.name.toString(), "APPLES");
             assert.deepStrictEqual(APPLES.description, null);
-            assert.deepStrictEqual(APPLES.span, new language.Span(50, 31));
+            assert.deepStrictEqual(APPLES.span, new Language.Span(50, 31));
 
             const apples: ParameterDefinition = dt.getParameterDefinition("'APPles'");
             assert(apples);
             assert.deepStrictEqual(apples.name.toString(), "apples");
             assert.deepStrictEqual(apples.description, null);
-            assert.deepStrictEqual(apples.span, new language.Span(18, 30));
+            assert.deepStrictEqual(apples.span, new Language.Span(18, 30));
         });
     });
 
@@ -596,13 +596,13 @@ suite("DeploymentTemplate", () => {
             assert(match0);
             assert.deepStrictEqual(match0.name.toString(), "apples");
             assert.deepStrictEqual(match0.description, null);
-            assert.deepStrictEqual(match0.span, new language.Span(18, 30));
+            assert.deepStrictEqual(match0.span, new Language.Span(18, 30));
 
             const match1: ParameterDefinition = matches[1];
             assert(match1);
             assert.deepStrictEqual(match1.name.toString(), "bananas");
             assert.deepStrictEqual(match1.description, null);
-            assert.deepStrictEqual(match1.span, new language.Span(50, 32));
+            assert.deepStrictEqual(match1.span, new Language.Span(50, 32));
         });
 
         test("with prefix of one of the parameters", () => {
@@ -616,7 +616,7 @@ suite("DeploymentTemplate", () => {
             assert(match0);
             assert.deepStrictEqual(match0.name.toString(), "apples");
             assert.deepStrictEqual(match0.description, null);
-            assert.deepStrictEqual(match0.span, new language.Span(18, 30));
+            assert.deepStrictEqual(match0.span, new Language.Span(18, 30));
         });
 
         test("with prefix of none of the parameters", () => {
@@ -635,7 +635,7 @@ suite("DeploymentTemplate", () => {
             assert(match0);
             assert.deepStrictEqual(match0.name.toString(), "apples");
             assert.deepStrictEqual(match0.description, null);
-            assert.deepStrictEqual(match0.span, new language.Span(18, 30));
+            assert.deepStrictEqual(match0.span, new Language.Span(18, 30));
         });
 
         test("with case sensitive and insensitive match", () => {
@@ -649,13 +649,13 @@ suite("DeploymentTemplate", () => {
             assert(match0);
             assert.deepStrictEqual(match0.name.toString(), "apples");
             assert.deepStrictEqual(match0.description, null);
-            assert.deepStrictEqual(match0.span, new language.Span(18, 30));
+            assert.deepStrictEqual(match0.span, new Language.Span(18, 30));
 
             const match1: ParameterDefinition = matches[1];
             assert(match1);
             assert.deepStrictEqual(match1.name.toString(), "APPLES");
             assert.deepStrictEqual(match1.description, null);
-            assert.deepStrictEqual(match1.span, new language.Span(50, 31));
+            assert.deepStrictEqual(match1.span, new Language.Span(50, 31));
         });
     });
 
@@ -704,7 +704,7 @@ suite("DeploymentTemplate", () => {
 
             const value: Json.Value = Json.asStringValue(apples.value);
             assert(value);
-            assert.deepStrictEqual(value.span, new language.Span(27, 5));
+            assert.deepStrictEqual(value.span, new Language.Span(27, 5));
             assert.deepStrictEqual(value.toString(), "yum");
         });
 
@@ -717,7 +717,7 @@ suite("DeploymentTemplate", () => {
 
             const value: Json.StringValue = Json.asStringValue(apples.value);
             assert(value);
-            assert.deepStrictEqual(value.span, new language.Span(27, 5));
+            assert.deepStrictEqual(value.span, new Language.Span(27, 5));
             assert.deepStrictEqual(value.toString(), "yum");
         });
 
@@ -730,7 +730,7 @@ suite("DeploymentTemplate", () => {
 
             const value: Json.StringValue = Json.asStringValue(apples.value);
             assert(value);
-            assert.deepStrictEqual(value.span, new language.Span(27, 5));
+            assert.deepStrictEqual(value.span, new Language.Span(27, 5));
             assert.deepStrictEqual(value.toString(), "yum");
         });
 
@@ -743,7 +743,7 @@ suite("DeploymentTemplate", () => {
 
             const value: Json.StringValue = Json.asStringValue(apples.value);
             assert(value);
-            assert.deepStrictEqual(value.span, new language.Span(27, 5));
+            assert.deepStrictEqual(value.span, new Language.Span(27, 5));
             assert.deepStrictEqual(value.toString(), "yum");
         });
 
@@ -754,10 +754,10 @@ suite("DeploymentTemplate", () => {
             assert(APPLES);
             assert.deepStrictEqual(APPLES.name.toString(), "APPLES");
 
-            const APPLESvalue: Json.StringValue = Json.asStringValue(APPLES.value);
-            assert(APPLESvalue);
-            assert.deepStrictEqual(APPLESvalue.span, new language.Span(44, 6));
-            assert.deepStrictEqual(APPLESvalue.toString(), "good");
+            const applesValue: Json.StringValue = Json.asStringValue(APPLES.value);
+            assert(applesValue);
+            assert.deepStrictEqual(applesValue.span, new Language.Span(44, 6));
+            assert.deepStrictEqual(applesValue.toString(), "good");
 
             const apples: Json.Property = dt.getVariableDefinition("'APPles'");
             assert(apples);
@@ -765,7 +765,7 @@ suite("DeploymentTemplate", () => {
 
             const value: Json.StringValue = Json.asStringValue(apples.value);
             assert(value);
-            assert.deepStrictEqual(value.span, new language.Span(27, 5));
+            assert.deepStrictEqual(value.span, new Language.Span(27, 5));
             assert.deepStrictEqual(value.toString(), "yum");
         });
     });
@@ -791,13 +791,13 @@ suite("DeploymentTemplate", () => {
             assert.deepStrictEqual(apples.name.toString(), "apples");
             const applesValue: Json.StringValue = Json.asStringValue(apples.value);
             assert(applesValue);
-            assert.deepStrictEqual(applesValue.span, new language.Span(27, 8));
+            assert.deepStrictEqual(applesValue.span, new Language.Span(27, 8));
             assert.deepStrictEqual(applesValue.toString(), "APPLES");
 
             const bananas: Json.Property = definitions[1];
             assert.deepStrictEqual(bananas.name.toString(), "bananas");
             const bananasValue: Json.NumberValue = Json.asNumberValue(bananas.value);
-            assert.deepStrictEqual(bananasValue.span, new language.Span(48, 2));
+            assert.deepStrictEqual(bananasValue.span, new Language.Span(48, 2));
         });
 
         test("with prefix of one of the variables", () => {
@@ -810,7 +810,7 @@ suite("DeploymentTemplate", () => {
             assert.deepStrictEqual(apples.name.toString(), "apples");
             const applesValue: Json.StringValue = Json.asStringValue(apples.value);
             assert(applesValue);
-            assert.deepStrictEqual(applesValue.span, new language.Span(27, 8));
+            assert.deepStrictEqual(applesValue.span, new Language.Span(27, 8));
             assert.deepStrictEqual(applesValue.toString(), "APPLES");
         });
 
@@ -879,7 +879,7 @@ suite("DeploymentTemplate", () => {
             const list: Reference.List = dt.findReferences(Reference.ReferenceKind.Parameter, "pName");
             assert(list);
             assert.deepStrictEqual(list.kind, Reference.ReferenceKind.Parameter);
-            assert.deepStrictEqual(list.spans, [new language.Span(19, 5)]);
+            assert.deepStrictEqual(list.spans, [new Language.Span(19, 5)]);
         });
 
         test("with variable type and no matching variable definition", () => {
@@ -895,7 +895,7 @@ suite("DeploymentTemplate", () => {
             const list: Reference.List = dt.findReferences(Reference.ReferenceKind.Variable, "vName");
             assert(list);
             assert.deepStrictEqual(list.kind, Reference.ReferenceKind.Variable);
-            assert.deepStrictEqual(list.spans, [new language.Span(18, 5)]);
+            assert.deepStrictEqual(list.spans, [new Language.Span(18, 5)]);
         });
     });
 });
@@ -915,6 +915,19 @@ suite("ReferenceInVariableDefinitionJSONVisitor", () => {
             const visitor = new ReferenceInVariableDefinitionJSONVisitor(dt);
             assert.deepStrictEqual(visitor.referenceSpans, []);
         });
+
+        testDiagnostics(
+            "error: reference in variable definition",
+            {
+                "variables": {
+                    "a": "[reference('test')]"
+                },
+            },
+            {},
+            [
+                "Error: reference() cannot be invoked inside of a variable definition. (ARM Tools)",
+                "Warning: The variable 'a' is never used. (ARM Tools)"
+            ]);
     });
 
     suite("visitStringValue(Json.StringValue)", () => {
@@ -946,7 +959,7 @@ suite("ReferenceInVariableDefinitionJSONVisitor", () => {
             const tle: Json.StringValue = Json.asStringValue(variablesObject.getPropertyValue("a"));
 
             visitor.visitStringValue(tle);
-            assert.deepStrictEqual(visitor.referenceSpans, [new language.Span(24, 9)]);
+            assert.deepStrictEqual(visitor.referenceSpans, [new Language.Span(24, 9)]);
         });
 
         test("with TLE string with reference() call inside concat() call", () => {
@@ -957,7 +970,112 @@ suite("ReferenceInVariableDefinitionJSONVisitor", () => {
             const tle: Json.StringValue = Json.asStringValue(variablesObject.getPropertyValue("a"));
 
             visitor.visitStringValue(tle);
-            assert.deepStrictEqual(visitor.referenceSpans, [new language.Span(31, 9)]);
+            assert.deepStrictEqual(visitor.referenceSpans, [new Language.Span(31, 9)]);
         });
+    });
+});
+
+suite("Incomplete JSON shouldn't crash parse", function (this: ISuiteCallbackContext): void {
+    this.timeout(10000);
+
+    const template: string =
+        `{
+        "$schema": "http://schema.management.azure.com/schemas/2015-01-01/deploymentTemplate.json#",
+        "contentVersion": "1.0.0.0",
+        "parameters": {
+            "location": { "type": "string" },
+            "networkInterfaceName": {
+                "type": "string"
+            },
+        },
+        "variables": {
+            "vnetId": "[resourceId(resourceGroup().name,'Microsoft.Network/virtualNetworks', parameters('virtualNetworkName'))]",
+        },
+        "resources": [
+            {
+                "name": "[parameters('networkInterfaceName')]",
+                "type": "Microsoft.Network/networkInterfaces",
+                "apiVersion": "2018-10-01",
+                "location": "[parameters('location')]",
+                "dependsOn": [
+                    "[concat('Microsoft.Network/networkSecurityGroups/', parameters('networkSecurityGroupName'))]",
+                    "[concat('Microsoft.Network/virtualNetworks/', parameters('virtualNetworkName'))]",
+                    "[concat('Microsoft.Network/publicIpAddresses/', parameters('publicIpAddressName'))]"
+                ],
+                "properties": {
+                    "$test-commandToExecute": "[concat('cd /hub*/docker-compose; sudo docker-compose down -t 60; sudo -s source /set_hub_url.sh ', reference(parameters('publicIpName')).dnsSettings.fqdn, ';  sudo docker volume rm ''dockercompose_cert-volume''; sudo docker-compose up')]",
+                    "ipConfigurations": [
+                        {
+                            "name": "ipconfig1",
+                            "properties": {
+                                "subnet": {
+                                    "id": "[variables('subnetRef')]"
+                                },
+                                "privateIPAllocationMethod": "Dynamic",
+                                "publicIpAddress": {
+                                    "id": "[resourceId(resourceGroup().name, 'Microsoft.Network/publicIpAddresses', parameters('publicIpAddressName'))]"
+                                }
+                            }
+                        }
+                    ]
+                },
+                "tags": {}
+            }
+        ],
+        "outputs": {
+            "adminUsername": {
+                "type": "string",
+                "value": "[parameters('adminUsername')]"
+            }
+        }
+    }
+    `;
+
+    test("https://github.com/Microsoft/vscode-azurearmtools/issues/193", async () => {
+        // Just make sure nothing throws
+        let modifiedTemplate = template.replace('"type": "string"', '"type": string');
+        let dt = new DeploymentTemplate(modifiedTemplate, "id");
+        await dt.errors;
+    });
+
+    test("typing character by character", async () => {
+        // Just make sure nothing throws
+        for (let i = 0; i < template.length; ++i) {
+            let partialTemplate = template.slice(0, i);
+            let dt = new DeploymentTemplate(partialTemplate, "id");
+            await dt.errors;
+        }
+    });
+
+    test("typing backwards character by character", async () => {
+        // Just make sure nothing throws
+        for (let i = 0; i < template.length; ++i) {
+            let partialTemplate = template.slice(i);
+            let dt = new DeploymentTemplate(partialTemplate, "id");
+            await dt.errors;
+        }
+    });
+
+    test("Random modifications", async () => {
+        // Just make sure nothing throws
+        let modifiedTemplate: string = template;
+
+        for (let i = 0; i < 1000; ++i) {
+            if (modifiedTemplate.length > 0 && Math.random() < 0.5) {
+                // Delete some characters
+                let position = Math.random() * (modifiedTemplate.length - 1);
+                let length = Math.random() * Math.max(5, modifiedTemplate.length);
+                modifiedTemplate = modifiedTemplate.slice(position, position + length);
+            } else {
+                // Insert some characters
+                let position = Math.random() * modifiedTemplate.length;
+                let length = Math.random() * 5;
+                let s = randomBytes(length).toString();
+                modifiedTemplate = modifiedTemplate.slice(0, position) + s + modifiedTemplate.slice(position);
+            }
+
+            let dt = new DeploymentTemplate(modifiedTemplate, "id");
+            await dt.errors;
+        }
     });
 });
