@@ -6,10 +6,8 @@
 // tslint:disable:promise-function-async
 
 import * as assert from "assert";
-
-import { httpGet } from "../src/httpGet";
-
-import { networkTest } from "./networkTest";
+import { httpGet } from "../extension.bundle";
+import { networkTest } from "./networkTest.test";
 
 suite("HttpClient", () => {
     suite("get(string)", () => {
@@ -17,7 +15,7 @@ suite("HttpClient", () => {
             return httpGet("http://www.bing.com")
                 .then((content: string) => {
                     assert(content, "Content was undefined, null, or empty");
-                    assert(content.includes("Word Online"), "Content did not include the phrase 'Word Online'");
+                    assert(content.includes("Bing"), "Content did not include the phrase 'Bing'");
                 });
         });
 
@@ -25,7 +23,7 @@ suite("HttpClient", () => {
             return httpGet("www.bing.com")
                 .then((content: string) => {
                     assert(content, "Content was undefined, null, or empty");
-                    assert(content.includes("Word Online"), "Content did not include the phrase 'Word Online'");
+                    assert(content.includes("Bing"), "Content did not include the phrase 'Bing'");
                 });
         });
 
@@ -33,7 +31,7 @@ suite("HttpClient", () => {
             return httpGet("https://storageexplorer.com")
                 .then((content: string) => {
                     assert(content, "No content");
-                    assert(content.includes("Azure Storage Explorer"), "Doesn't include");
+                    assert(content.includes("Azure Storage Explorer"), "Doesn't include 'Azure Storage Explorer'");
                 });
         });
 
@@ -43,7 +41,7 @@ suite("HttpClient", () => {
                     assert(false, "Expected the catch function to be called.");
                 })
                 // tslint:disable-next-line:no-any
-                .catch((reason: any) => {
+                .catch((reason: { code?: string; errno?: number; hostname?: string; syscall?: string }) => {
                     assert(reason);
                     assert.deepStrictEqual(reason.code, "ENOTFOUND");
                     assert.deepStrictEqual(reason.errno, "ENOTFOUND");
