@@ -3,6 +3,8 @@
  *  Licensed under the MIT License. See LICENSE.md in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
+import * as path from 'path';
+
 //
 // PLEASE DO NOT MODIFY / DELETE UNLESS YOU KNOW WHAT YOU ARE DOING
 //
@@ -15,9 +17,10 @@
 // to report the results back to the caller. When the tests are finished, return
 // a possible error to the callback or null if none.
 
-import testRunner = require('vscode/lib/testrunner');
+// tslint:disable-next-line:no-require-imports no-var-requires no-submodule-imports
+let testRunner = require('vscode/lib/testrunner');
 
-const options: { [key: string]: string | boolean | number } = {
+let options: { [key: string]: string | boolean | number | object } = {
     ui: 'tdd', 		// the TDD UI is being used in extension.test.ts (suite, test, etc.)
     useColors: true // colored output from test results
 };
@@ -36,20 +39,32 @@ const options: { [key: string]: string | boolean | number } = {
 //
 // See https://github.com/mochajs/mocha/wiki/Using-mocha-programmatically#set-options for all available options
 
-for (const envVar of Object.keys(process.env)) {
-    const match: RegExpMatchArray | null = envVar.match(/^mocha_(.+)/i);
+// Defaults
+options.reporter = 'mocha-multi-reporters';
+options.reporterOptions = {
+    reporterEnabled: "spec, mocha-junit-reporter",
+    mochaJunitReporterReporterOptions: {
+        mochaFile: path.join(__dirname, '..', '..', 'test-results.xml')
+    }
+};
+
+let environmentVariables = <{ [key: string]: string }>process.env;
+for (let envVar of Object.keys(environmentVariables)) {
+    let match = envVar.match(/^mocha_(.+)/i);
     if (match) {
-        const [, option] = match;
-        // tslint:disable-next-line:strict-boolean-expressions
-        let value: string | number = process.env[envVar] || '';
-        if (!isNaN(parseInt(value))) {
-            value = parseInt(value);
+        let [, option] = match;
+        let value: string | number = environmentVariables[envVar];
+        if (typeof value === 'string' && !isNaN(parseInt(value, undefined))) {
+            value = parseInt(value, undefined);
         }
+
         options[option] = value;
     }
 }
+
 console.warn(`Mocha options: ${JSON.stringify(options, null, 2)}`);
 
+// tslint:disable-next-line: no-unsafe-any
 testRunner.configure(options);
 
 module.exports = testRunner;
